--- src/pages/admin/report/new/data.ts
@@ -1,22 +1,0 @@
-const participant = {
-  name: '',
-  studentId: '',
-  department: '',
-};
-
-export const EMPTY_DATA = {
-  term: 0,
-  date: { startDate: null, endDate: null },
-  place: '',
-  startTime: '',
-  endTime: '',
-  uploadFiles: null,
-  content: '',
-  participants: [
-    participant,
-    participant,
-    participant,
-    participant,
-    participant,
-  ],
-};
